Statically resolve calling of known function objects
This can happen in cases like:
d = {}
im = d.__setitem__
im(1, 2)

We will know that im is an instancemethod over d and the function dictSetItem,
but we previously didn't know how to apply dictSetItem so we would just fall
back on the generic runtimeCall behavior.

We already had this functionality when going through a callattr route, so
just split that up and separate out the function-calling part.  Then,
detect if we have a constant BoxedFunction object, and use that static-function-
calling code.

literal_perf.py is still failing since we don't fast-path hash