Trying another approach: leaving UNBOXED_INT as the old INT
seems to be going ok, but running into something that I think might be unfixable.
all the unboxed int code is left in UNBOXED_INT, but it needs to know whether it
is part of a [boxable] INT or not.  standalone UNBOXED_INT operations don't really
make sense themselves.

I think there's a slightly different distinction being drawn here.  i64 is a "concrete"
type that can be passed around.  it's not a "usable" type though.  for now usable means
kind of means "boxed".  box-unbox-pair-int is not "concrete", but it is "usable".