C++ refcounting type helper
The tricky part is getting it to check refcounting discipline.  The easy
way to do this is to make the ctor incref and the dtor decref, but this
has a couple of perf issues:
- It introduces extra incref/decref operations (you can't pass an owned
  reference without an extra incref/decref pair)
- It makes it no longer a value type.

Maybe there's a good solution this way, but I couldn't find it.  So go with
a different approach, of keeping some extra bookkeeping state and doing
asserts on that if a flag is turned on.

Other things that are tricky:
- even if we want checking turned on most places, we probably want the ability
  specifically request no checking (ex in data structures).
- we want these to respect the type system

WIP!
- this kind of works, but needs lots of extra annotations: "int_cls.init(owned(new BoxedClass()))"
- auto-refcounting is nice, but will screw with CAPI code.